@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { CrudService } from '../shared/crud.service';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 import { ToastrService } from 'ngx-toastr';
 
 
@@ -11,11 +11,11 @@ import { ToastrService } from 'ngx-toastr';
 })
 
 export class AddStudentComponent implements OnInit {
-  public studentForm: FormGroup;
+  public studentForm: UntypedFormGroup;
  
   constructor(
     public crudApi: CrudService,
-    public fb: FormBuilder,
+    public fb: UntypedFormBuilder,
     public toastr: ToastrService
   ) { }
 
